chartgenius-v3/backend/requirements.txt
# ChartGenius v3 Backend Dependencies
# Updated for 2025 with modern versions

# Core FastAPI and ASGI
fastapi>=0.115.14
uvicorn[standard]>=0.34.1
pydantic>=2.11.3

# HTTP Client
httpx>=0.28.1

# Data Processing
pandas>=2.2.3
numpy>=2.1.3
ta>=0.11.0
statsmodels>=0.14.4
scipy>=1.14.1

# AI/LLM Integration
openai>=1.14.0
google-generativeai>=0.7.0

# Authentication & Security
PyJWT>=2.10.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
oracledb>=3.2.0

# Caching
redis>=5.0.0

# Telegram Integration
python-telegram-bot>=21.11.1
telegram-webapp-auth>=3.0.5

# Payments (Telegram Stars)
stripe>=7.0.0

# Utilities
python-dotenv>=1.0.1
pyyaml>=6.0.2
ulid-py>=1.1.0
jinja2>=3.1.4

# Visualization (if needed)
plotly>=5.24.1

# Development & Testing
pytest>=8.3.4
pytest-asyncio>=0.24.0
httpx>=0.28.1

# Monitoring & Logging
structlog>=24.4.0
